+++ depthai_ros_driver/include/depthai_ros_driver/dai_nodes/nn/detection.hpp
@@ -4,7 +4,7 @@
 #include <string>
 #include <vector>
 
-#include "camera_info_manager/camera_info_manager.hpp"
+#include "camera_info_manager/camera_info_manager.h"
 #include "depthai/device/DataQueue.hpp"
 #include "depthai/device/Device.hpp"
 #include "depthai/pipeline/Pipeline.hpp"
@@ -16,9 +16,10 @@
 #include "depthai_ros_driver/dai_nodes/base_node.hpp"
 #include "depthai_ros_driver/dai_nodes/sensors/sensor_helpers.hpp"
 #include "depthai_ros_driver/param_handlers/nn_param_handler.hpp"
-#include "image_transport/camera_publisher.hpp"
-#include "image_transport/image_transport.hpp"
-#include "rclcpp/node.hpp"
+#include "depthai_ros_driver/parametersConfig.h"
+#include "image_transport/camera_publisher.h"
+#include "image_transport/image_transport.h"
+#include "ros/node_handle.h"
 
 namespace depthai_ros_driver {
 
@@ -35,14 +36,14 @@ class Detection : public BaseNode {
      * @param      node         The node
      * @param      pipeline     The pipeline
      */
-    Detection(const std::string& daiNodeName, rclcpp::Node* node, std::shared_ptr<dai::Pipeline> pipeline) : BaseNode(daiNodeName, node, pipeline) {
-        RCLCPP_DEBUG(node->get_logger(), "Creating node %s", daiNodeName.c_str());
+    Detection(const std::string& daiNodeName, ros::NodeHandle node, std::shared_ptr<dai::Pipeline> pipeline) : BaseNode(daiNodeName, node, pipeline), it(node) {
+        ROS_DEBUG("Creating node %s", daiNodeName.c_str());
         setNames();
         detectionNode = pipeline->create<T>();
         imageManip = pipeline->create<dai::node::ImageManip>();
         ph = std::make_unique<param_handlers::NNParamHandler>(node, daiNodeName);
         ph->declareParams(detectionNode, imageManip);
-        RCLCPP_DEBUG(node->get_logger(), "Node %s created", daiNodeName.c_str());
+        ROS_DEBUG("Node %s created", daiNodeName.c_str());
         imageManip->out.link(detectionNode->input);
         setXinXout(pipeline);
     }
@@ -59,8 +60,8 @@ class Detection : public BaseNode {
         int width;
         int height;
         if(ph->getParam<bool>("i_disable_resize")) {
-            width = getROSNode()->get_parameter("rgb.i_preview_size").as_int();
-            height = getROSNode()->get_parameter("rgb.i_preview_size").as_int();
+            width = ph->getOtherNodeParam<int>("rgb", "i_preview_size");
+            height = ph->getOtherNodeParam<int>("rgb", "i_preview_size");
         } else {
             width = imageManip->initialConfig.getResizeConfig().width;
             height = imageManip->initialConfig.getResizeConfig().height;
@@ -68,25 +69,17 @@ class Detection : public BaseNode {
         detConverter = std::make_unique<dai::ros::ImgDetectionConverter>(
             tfPrefix + "_camera_optical_frame", width, height, false, ph->getParam<bool>("i_get_base_device_timestamp"));
         detConverter->setUpdateRosBaseTimeOnToRosMsg(ph->getParam<bool>("i_update_ros_base_time_on_ros_msg"));
-        rclcpp::PublisherOptions options;
-        options.qos_overriding_options = rclcpp::QosOverridingOptions();
-        detPub = getROSNode()->template create_publisher<vision_msgs::msg::Detection2DArray>("~/" + getName() + "/detections", 10, options);
+        detPub = getROSNode().template advertise<vision_msgs::Detection2DArray>(getName() + "/detections", 10);
         nnQ->addCallback(std::bind(&Detection::detectionCB, this, std::placeholders::_1, std::placeholders::_2));
 
         if(ph->getParam<bool>("i_enable_passthrough")) {
             ptQ = device->getOutputQueue(ptQName, ph->getParam<int>("i_max_q_size"), false);
             imageConverter = std::make_unique<dai::ros::ImageConverter>(tfPrefix + "_camera_optical_frame", false);
             imageConverter->setUpdateRosBaseTimeOnToRosMsg(ph->getParam<bool>("i_update_ros_base_time_on_ros_msg"));
-            infoManager = std::make_shared<camera_info_manager::CameraInfoManager>(
-                getROSNode()->create_sub_node(std::string(getROSNode()->get_name()) + "/" + getName()).get(), "/" + getName());
-            infoManager->setCameraInfo(sensor_helpers::getCalibInfo(getROSNode()->get_logger(),
-                                                                    *imageConverter,
-                                                                    device,
-                                                                    static_cast<dai::CameraBoardSocket>(ph->getParam<int>("i_board_socket_id")),
-                                                                    width,
-                                                                    height));
+            infoManager = std::make_shared<camera_info_manager::CameraInfoManager>(ros::NodeHandle(getROSNode(), getName()), "/" + getName());
+            infoManager->setCameraInfo(sensor_helpers::getCalibInfo(*imageConverter, device, dai::CameraBoardSocket::CAM_A, width, height));
 
-            ptPub = image_transport::create_camera_publisher(getROSNode(), "~/" + getName() + "/passthrough/image_raw");
+            ptPub = it.advertiseCamera(getName() + "/passthrough/image_raw", 1);
             ptQ->addCallback(std::bind(sensor_helpers::basicCameraPub, std::placeholders::_1, std::placeholders::_2, *imageConverter, ptPub, infoManager));
         }
     };
@@ -112,7 +105,6 @@ class Detection : public BaseNode {
         }
         return imageManip->inputImage;
     };
-
     void setNames() override {
         nnQName = getName() + "_nn";
         ptQName = getName() + "_pt";
@@ -142,8 +134,8 @@ class Detection : public BaseNode {
         }
     };
 
-    void updateParams(const std::vector<rclcpp::Parameter>& params) override {
-        ph->setRuntimeParams(params);
+    void updateParams(parametersConfig& config) override {
+        ph->setRuntimeParams(config);
     };
 
    private:
@@ -155,17 +147,18 @@ class Detection : public BaseNode {
      */
     void detectionCB(const std::string& /*name*/, const std::shared_ptr<dai::ADatatype>& data) {
         auto inDet = std::dynamic_pointer_cast<dai::ImgDetections>(data);
-        std::deque<vision_msgs::msg::Detection2DArray> deq;
+        std::deque<vision_msgs::Detection2DArray> deq;
         detConverter->toRosMsg(inDet, deq);
         while(deq.size() > 0) {
             auto currMsg = deq.front();
-            detPub->publish(currMsg);
+            detPub.publish(currMsg);
             deq.pop_front();
         }
     };
     std::unique_ptr<dai::ros::ImgDetectionConverter> detConverter;
+    image_transport::ImageTransport it;
     std::vector<std::string> labelNames;
-    rclcpp::Publisher<vision_msgs::msg::Detection2DArray>::SharedPtr detPub;
+    ros::Publisher detPub;
     std::unique_ptr<dai::ros::ImageConverter> imageConverter;
     image_transport::CameraPublisher ptPub;
     std::shared_ptr<camera_info_manager::CameraInfoManager> infoManager;