+++ depthai-ros/CHANGELOG.rst
@@ -2,6 +2,15 @@
 Changelog for package depthai-ros
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+2.8.2 (2023-10-17)
+-------------------
+
+* Fixed default resolution for Stereo cameras
+* Added CameraInfo update based on alpha scaling
+* Logger restart bugfix
+* URDF parameters fix
+
+
 2.8.1 (2023-09-12)
 -------------------
 
@@ -10,6 +19,7 @@ Changelog for package depthai-ros
 * Updated sensor socket logic
 * Fixed issues when running robot_state_publisher as component
 * Added missing tf2 dependencies
+
 
 2.8.0 (2023-09-01)
 -------------------
@@ -42,7 +52,7 @@ Changelog for package depthai-ros
 * Camera trigger mechanisms
 * Brightness filter
 
-2.7.2 (2023-05-08)
+2.7.2 (2023-5-08)
 -------------------
 * IMU improvements